Migrate upsert_magic().
